+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT997DZJXKD7D6H0E233VCGK"
+        "01KT9EQSHWA3KCHWFEMKJE3X88"
     )
     r"""The unique identifier of the data source"""