+++ packages/orq-rc/src/orq_ai_sdk/models/createeval_guardrail_config_evals_response_200_applicationjson_responsebody_4_type.py
@@ -1192,9 +1192,9 @@ class Typescript(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-04T12:19:57.790Z"
+    created: Optional[str] = "2026-06-04T13:56:08.431Z"
 
-    updated: Optional[str] = "2026-06-04T12:19:57.790Z"
+    updated: Optional[str] = "2026-06-04T13:56:08.431Z"
 
     updated_by_id: OptionalNullable[str] = UNSET
 
@@ -1449,9 +1449,9 @@ class Ragas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-06-04T12:19:57.790Z"
+    created: Optional[str] = "2026-06-04T13:56:08.431Z"
 
-    updated: Optional[str] = "2026-06-04T12:19:57.790Z"
+    updated: Optional[str] = "2026-06-04T13:56:08.431Z"
 
     updated_by_id: OptionalNullable[str] = UNSET
 
@@ -2195,9 +2195,9 @@ class CreateEvalResponseBodyFunction(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-04T12:19:57.790Z"
+    created: Optional[str] = "2026-06-04T13:56:08.431Z"
 
-    updated: Optional[str] = "2026-06-04T12:19:57.790Z"
+    updated: Optional[str] = "2026-06-04T13:56:08.431Z"
 
     updated_by_id: OptionalNullable[str] = UNSET