+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT997DZEKB74HR2ARTSSAXKY"
+        "01KT9EQSHQG2M4BAG8E8RCVP90"
     )
     r"""The unique identifier of the data source"""