+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT997DZGXHCEYWWWS1RZ9AZV"
+        "01KT9EQSHT12B69P94MTJ2XHD7"
     )
     r"""The unique identifier of the data source"""