+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT997DZN05HN4NT2WXZPZTMC"
+        "01KT9EQSJ0PFGDQ1X4AXTNKTYF"
     )
     r"""The unique identifier of the data source"""